Squashed commit of the following:
commit 1941aab6b7d1521e1b26b35d39a67daa7838b911
Author: MasterlyMisheirt <thewizardofmich@gmail.com>
Date:   Wed Mar 19 19:51:33 2025 +0000

    Bottom Sheets for Activity and User

+++ app/src/main/java/com/example/themadproject/model/data/User.kt
@@ -9,6 +9,6 @@ data class User(
     val UserLongitude: Double,
     val UserPassword: String,
     val UserPhone: String,
-    val UserTimestamp: Int,
+    val UserTimestamp: Double,
     val UserUsername: String
 )
